Move old Ansible-config to backup directory
I just want to recreate all the Ansible configuration, to make sure everything is
up to date and in order. Last time I didn't know how to create Ansible-playbooks
and roles, so I might have messed it up a bit. So it's better to recreate
everything.

I could just have deleted the old stuff, but I might want to have a peek of what
I did last time. And it's easier to look at the directory than the Git log (yeah,
I'm a bit lazy).