Add 127.0.0.1 to apiserver cert
I thought that I might have to add 127 to the IP-adresses for the
certificate in order to be able to connect to Etcd from localhost.
But I was wrong, I had to connect thru the public IP-adress. This
might have been a configuration issue (but I leave the IP in there
any ways).