Only create certificate if it does not exists
If we only just recreate every cert when we run the script the CA cert will
be overwritten everytime. Which makes us update the CA for every
node (and restarting everything) everytime we run the script.

Add an if-case for every cert for every cert just to be on the safe side.